feat(query): refactor useLogout and useMyInfo hooks to utilize queryFns for better query management
