+++ apps/web/@hook/useMyInfo.ts
@@ -6,7 +6,7 @@ import { AuthUser } from "@repo/types";
 import { fetchMe } from "../@utils/query/user";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { updateMe } from "../@utils/query/user";
-import { queryKeys } from "../@utils/query/query.keys";
+import { queryFns, queryKeys } from "../@utils/query/query.control";
 
 export const useGetMyInfo = () => {
   const {
@@ -16,7 +16,7 @@ export const useGetMyInfo = () => {
     refetch,
   } = useQuery<AuthUser>({
     queryKey: queryKeys.user.base,
-    queryFn: fetchMe,
+    queryFn: queryFns.user.getMe,
     retry: false, // 로그인 실패 시 재시도 없음
   });
 
@@ -27,7 +27,7 @@ export const useUpdateMyInfo = () => {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: updateMe,
+    mutationFn: queryFns.user.updateMe,
     onSuccess: (updatedUser) => {
       // 'auth/me' 쿼리 캐시 갱신
       queryClient.setQueryData(queryKeys.user.base, updatedUser);
